+++ demo/urls.py
@@ -13,14 +13,22 @@ Including another URLconf
     1. Import the include() function: from django.urls import include, path
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
-# from django.contrib import admin
 from django.urls import path, include
 from django.conf import settings
 
 urlpatterns = [
-    # path('admin/', admin.site.urls),
-    path("auth/", include("djplus.auth.urls", namespace="auth")),
+
 ]
+
+if "django.contrib.admin" in settings.INSTALLED_APPS:
+    from django.contrib import admin
+    urlpatterns += [path("admin/", admin.site.urls)]
+
+if "djplus.auth" in settings.INSTALLED_APPS:
+    urlpatterns += [path("auth/", include("djplus.auth.urls", namespace="auth"))]
+
+if "djplus.blog" in settings.INSTALLED_APPS:
+    urlpatterns += [path("blog/", include("djplus.blog.urls", namespace="blog"))]
 
 if "debug_toolbar" in settings.INSTALLED_APPS:
     urlpatterns += [path("__debug__/", include("debug_toolbar.urls", namespace="djdt"))]