+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- add initial app called `blog`.
+- add `expire_date` and `data` fields to `Session` model.
+### Changed
+- Use `djplus.toml` file instead of `djplus.json` file to manage configurations.
+
 
 ## [2.0.0] - 2022-07-25
 ### Added